+++ openbank-ledger-service/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.20.1](https://github.com/JiRaska/open-bank-oss/compare/ledger-service-v1.20.0...ledger-service-v1.20.1) (2026-08-09)
+
+
+### Bug Fixes
+
+* **ledger:** make the ČNB fixing's age observable at the point of revaluation ([#4219](https://github.com/JiRaska/open-bank-oss/issues/4219)) ([368b2fe](https://github.com/JiRaska/open-bank-oss/commit/368b2fe33f98b28aaba07305df0e035fc0243cc2))
+
 ## [1.20.0](https://github.com/JiRaska/open-bank-oss/compare/ledger-service-v1.19.0...ledger-service-v1.20.0) (2026-08-07)